WIP: FSDEX-305: Migrate LQF minting to OPTIMISM

+++ contracts/LiquifiInitialGovernor.sol
@@ -1,13 +1,14 @@
 // SPDX-License-Identifier: GPL-3.0
-pragma solidity >= 0.7.0 <0.8.0;
+pragma solidity >=0.7.0 <0.8.0;
 
 import {LiquifiProposal} from "./LiquifiProposal.sol";
 import {LiquifiDAO} from "./libraries/LiquifiDAO.sol";
 import {GovernanceRouter} from "./interfaces/GovernanceRouter.sol";
 import {ERC20} from "./interfaces/ERC20.sol";
-import { DelayedExchangePool } from "./interfaces/DelayedExchangePool.sol";
-import { Liquifi } from "./libraries/Liquifi.sol";
-import { Math } from "./libraries/Math.sol";
+import {DelayedExchangePool} from "./interfaces/DelayedExchangePool.sol";
+import {Liquifi} from "./libraries/Liquifi.sol";
+import {Math} from "./libraries/Math.sol";
+
 //import { Debug } from "./libraries/Debug.sol";
 
 contract LiquifiInitialGovernor {
@@ -16,65 +17,74 @@ contract LiquifiInitialGovernor {
     event EmergencyLock(address sender, address pool);
     event ProposalCreated(address proposal);
     event ProposalFinalized(address proposal, LiquifiDAO.ProposalStatus proposalStatus);
-    event DepositWithdrawn(address user, uint amount);
+    event DepositWithdrawn(address user, uint256 amount);
 
-    struct CreatedProposals{
-        uint amountDeposited;
+    struct CreatedProposals {
+        uint256 amountDeposited;
         LiquifiDAO.ProposalStatus status;
         address creator;
     }
 
     struct Deposit {
-        uint amount;
-        uint unfreezeTime;
+        uint256 amount;
+        uint256 unfreezeTime;
     }
-    
+
     LiquifiProposal[] public deployedProposals;
     mapping(address => CreatedProposals) proposalInfo;
-    mapping(/* user */address => Deposit) public deposits;
+    /* user */
+    mapping(address => Deposit) public deposits;
     address[] public userDepositsList;
 
-    uint public immutable tokensRequiredToCreateProposal; 
-    uint public constant quorum = 10; //percenrage
-    uint public constant threshold = 50;
-    uint public constant vetoPercentage = 33;
-    uint public immutable votingPeriod; //hours
+    uint256 public immutable tokensRequiredToCreateProposal;
+    uint256 public constant quorum = 10; //percenrage
+    uint256 public constant threshold = 50;
+    uint256 public constant vetoPercentage = 33;
+    uint256 public immutable votingPeriod; //hours
 
     ERC20 private immutable govToken;
     GovernanceRouter public immutable governanceRouter;
 
-    constructor(address _governanceRouterAddress, uint _tokensRequiredToCreateProposal, uint _votingPeriod) {
+    constructor(
+        address _governanceRouterAddress,
+        uint256 _tokensRequiredToCreateProposal,
+        uint256 _votingPeriod
+    ) {
         tokensRequiredToCreateProposal = _tokensRequiredToCreateProposal;
         votingPeriod = _votingPeriod;
         govToken = GovernanceRouter(_governanceRouterAddress).minter();
         governanceRouter = GovernanceRouter(_governanceRouterAddress);
-        (address oldGovernor,) = GovernanceRouter(_governanceRouterAddress).governance();
+        (address oldGovernor, ) = GovernanceRouter(_governanceRouterAddress).governance();
         if (oldGovernor == address(0)) {
             GovernanceRouter(_governanceRouterAddress).setGovernor(address(this));
         }
     }
 
-    function deposit(address user, uint amount, uint unfreezeTime) private {
-        uint deposited = deposits[user].amount;
+    function deposit(
+        address user,
+        uint256 amount,
+        uint256 unfreezeTime
+    ) private {
+        uint256 deposited = deposits[user].amount;
         if (deposited < amount) {
-            uint remainingAmount = amount.subWithClip(deposited);
+            uint256 remainingAmount = amount.subWithClip(deposited);
             require(govToken.transferFrom(user, address(this), remainingAmount), "LIQUIFI_GV: TRANSFER FAILED");
             deposits[user].amount = amount;
         }
         deposits[user].unfreezeTime = Math.max(deposits[user].unfreezeTime, unfreezeTime);
         userDepositsList.push(user);
-    } 
+    }
 
     function withdraw() public {
         require(_withdraw(msg.sender, block.timestamp) > 0, "LIQUIFI_GV: WITHDRAW FAILED");
     }
 
-    function _withdraw(address user, uint maxTime) private returns (uint) {
-        uint amount = deposits[user].amount;
+    function _withdraw(address user, uint256 maxTime) private returns (uint256) {
+        uint256 amount = deposits[user].amount;
         if (amount == 0 || deposits[user].unfreezeTime >= maxTime) {
             return 0;
         }
-        
+
         deposits[user].amount = 0;
         require(govToken.transfer(user, amount), "LIQUIFI_GV: TRANSFER FAILED");
         emit DepositWithdrawn(user, amount);
@@ -85,24 +95,43 @@ contract LiquifiInitialGovernor {
         withdrawMultiple(0, userDepositsList.length);
     }
 
-    function withdrawMultiple(uint fromIndex, uint toIndex) public {
-        uint maxWithdrawTime = block.timestamp;
-        (address currentGovernor,) = governanceRouter.governance();
+    function withdrawMultiple(uint256 fromIndex, uint256 toIndex) public {
+        uint256 maxWithdrawTime = block.timestamp;
+        (address currentGovernor, ) = governanceRouter.governance();
 
         if (currentGovernor != address(this)) {
-            maxWithdrawTime = type(uint).max;
+            maxWithdrawTime = type(uint256).max;
         }
-        
-        for(uint userIndex = fromIndex; userIndex < toIndex; userIndex++) {
+
+        for (uint256 userIndex = fromIndex; userIndex < toIndex; userIndex++) {
             _withdraw(userDepositsList[userIndex], maxWithdrawTime);
         }
     }
 
-    function createProposal(string memory _proposal, uint _option, uint _newValue, address _address, address _address2) public {
+    function createProposal(
+        string memory _proposal,
+        uint256 _option,
+        uint256 _newValue,
+        address _address,
+        address _address2
+    ) public {
         address creator = msg.sender;
-        LiquifiProposal newProposal = new LiquifiProposal(_proposal, govToken.totalSupply(), address(govToken), _option, _newValue, quorum, threshold, vetoPercentage, votingPeriod, _address, _address2);
-        
-        uint tokensRequired = deposits[creator].amount.add(tokensRequiredToCreateProposal);
+        LiquifiProposal newProposal =
+            new LiquifiProposal(
+                _proposal,
+                govToken.totalSupply(),
+                address(govToken),
+                _option,
+                _newValue,
+                quorum,
+                threshold,
+                vetoPercentage,
+                votingPeriod,
+                _address,
+                _address2
+            );
+
+        uint256 tokensRequired = deposits[creator].amount.add(tokensRequiredToCreateProposal);
         deposit(creator, tokensRequired, newProposal.endTime());
 
         deployedProposals.push(newProposal);
@@ -113,12 +142,14 @@ contract LiquifiInitialGovernor {
     }
 
     function emergencyLock(address pool) public returns (bool locked) {
-        uint gasBefore = gasleft();
+        uint256 gasBefore = gasleft();
         try DelayedExchangePool(pool).processDelayedOrders() {
             return false;
-        } catch (bytes memory /*lowLevelData*/) {
-            uint gasAfter = gasleft();
-            require((gasBefore - gasAfter) * 10 / gasBefore >= 1, "LIQUIFI: LOW GAS");
+        } catch (
+            bytes memory /*lowLevelData*/
+        ) {
+            uint256 gasAfter = gasleft();
+            require(((gasBefore - gasAfter) * 10) / gasBefore >= 1, "LIQUIFI: LOW GAS");
             lockPool(pool);
             if (knownPool(pool)) {
                 emit EmergencyLock(msg.sender, pool);
@@ -131,27 +162,43 @@ contract LiquifiInitialGovernor {
         return deployedProposals;
     }
 
-    function proposalVote(address user, uint influence, uint unfreezeTime) public {
+    function proposalVote(
+        address user,
+        uint256 influence,
+        uint256 unfreezeTime
+    ) public {
         address proposal = msg.sender;
         require(proposalInfo[proposal].amountDeposited > 0, "LIQUIFI_GV: BAD SENDER");
-        require(proposalInfo[proposal].status == LiquifiDAO.ProposalStatus.IN_PROGRESS, "LIQUIFI_GV: PROPOSAL FINALIZED");
+        require(
+            proposalInfo[proposal].status == LiquifiDAO.ProposalStatus.IN_PROGRESS,
+            "LIQUIFI_GV: PROPOSAL FINALIZED"
+        );
 
         deposit(user, influence, unfreezeTime);
     }
 
-    function proposalFinalization(LiquifiDAO.ProposalStatus _proposalStatus, uint _option, uint /* _value */, address _address, address /* _address2 */) public {
+    function proposalFinalization(
+        LiquifiDAO.ProposalStatus _proposalStatus,
+        uint256 _option,
+        uint256, /* _value */
+        address _address,
+        address /* _address2 */
+    ) public {
         address proposal = msg.sender;
         require(proposalInfo[proposal].amountDeposited > 0, "LIQUIFI_GV: BAD SENDER");
-        require(proposalInfo[proposal].status == LiquifiDAO.ProposalStatus.IN_PROGRESS, "LIQUIFI_GV: PROPOSAL FINALIZED");
-        
+        require(
+            proposalInfo[proposal].status == LiquifiDAO.ProposalStatus.IN_PROGRESS,
+            "LIQUIFI_GV: PROPOSAL FINALIZED"
+        );
+
         if (_proposalStatus == LiquifiDAO.ProposalStatus.APPROVED) {
-            if (_option == 1) { 
-                changeGovernor(_address); 
+            if (_option == 1) {
+                changeGovernor(_address);
             }
         }
 
-        proposalInfo[proposal].status = _proposalStatus;   
-        emit ProposalFinalized(proposal, _proposalStatus);   
+        proposalInfo[proposal].status = _proposalStatus;
+        emit ProposalFinalized(proposal, _proposalStatus);
     }
 
     function changeGovernor(address _newGovernor) private {
@@ -159,10 +206,10 @@ contract LiquifiInitialGovernor {
     }
 
     function lockPool(address pool) internal {
-        (,uint governancePacked,,,,,,) = DelayedExchangePool(pool).poolState();
+        (, uint256 governancePacked, , , , , , ) = DelayedExchangePool(pool).poolState();
 
-        governancePacked = governancePacked | (1 << uint(Liquifi.Flag.POOL_LOCKED));
-        governancePacked = governancePacked | (1 << uint(Liquifi.Flag.GOVERNANCE_OVERRIDEN));
+        governancePacked = governancePacked | (1 << uint256(Liquifi.Flag.POOL_LOCKED));
+        governancePacked = governancePacked | (1 << uint256(Liquifi.Flag.GOVERNANCE_OVERRIDEN));
         DelayedExchangePool(pool).applyGovernance(governancePacked);
     }
