+++ contracts/interfaces/LiquifiCallee.sol
@@ -1,8 +1,14 @@
 // SPDX-License-Identifier: GPL-3.0
-pragma solidity >= 0.7.0 <0.8.0;
+pragma solidity >=0.7.0 <0.8.0;
 
 interface LiquifiCallee {
     // availableBalance contains 128 bits of availableBalanceA and 128 bits of availableBalanceB
     // delayedSwapsIncome contains 128 bits of delayedSwapsIncomeA and 128 bits of delayedSwapsIncomeB
-    function onLiquifiSwap(bytes calldata data, address sender, uint availableBalance, uint delayedSwapsIncome, uint instantNotFee) external;
+    function onLiquifiSwap(
+        bytes calldata data,
+        address sender,
+        uint256 availableBalance,
+        uint256 delayedSwapsIncome,
+        uint256 instantNotFee
+    ) external;
 }
