+++ contracts/interfaces/PoolFactory.sol
@@ -1,11 +1,14 @@
 // SPDX-License-Identifier: GPL-3.0
-pragma solidity >= 0.7.0 <0.8.0;
+pragma solidity >=0.7.0 <0.8.0;
 
 interface PoolFactory {
     event PoolCreatedEvent(address tokenA, address tokenB, bool aIsWETH, address indexed pool);
 
     function getPool(address tokenA, address tokenB) external returns (address);
+
     function findPool(address tokenA, address tokenB) external view returns (address);
-    function pools(uint poolIndex) external view returns (address pool);
-    function getPoolCount() external view returns (uint);
+
+    function pools(uint256 poolIndex) external view returns (address pool);
+
+    function getPoolCount() external view returns (uint256);
 }
