+++ contracts/interfaces/WETH.sol
@@ -1,9 +1,10 @@
 // SPDX-License-Identifier: GPL-3.0
-pragma solidity >= 0.7.0 <0.8.0;
+pragma solidity >=0.7.0 <0.8.0;
 
-import { ERC20 } from "./ERC20.sol";
+import {ERC20} from "./ERC20.sol";
 
 interface WETH is ERC20 {
     function deposit() external payable;
-    function withdraw(uint) external;
+
+    function withdraw(uint256) external;
 }
